+++ eufy_event_gateway/src/provider/devices/types.ts
@@ -26,6 +26,7 @@ export type DeviceCapability =
   | "battery"
   | "enableControl"
   | "motionDetectionControl"
+  | "sirenControl"
   | "privacyMode"
   | "lockControl";